+++ app/code/local/Stuntcoders/Slideshow/Helper/Data.php
@@ -16,4 +16,30 @@ class Stuntcoders_Slideshow_Helper_Data extends Mage_Core_Helper_Abstract
     {
         return Mage::getBaseUrl('media') . '/stuntcoders/slideshow/' . $imageName;
     }
+
+    public function generateConfig($data)
+    {
+        $config = array(
+            'adaptiveHeight' => true,
+            'infinite' => true,
+        );
+
+        if ($data['autoplay']) {
+            $config['autoplay'] = true;
+        }
+
+        if ($data['animation']) {
+            $config['fade'] = true;
+        }
+
+        if ($data['direction']) {
+            $config['vertical'] = true;
+        }
+
+        if ($data['dots']) {
+            $config['dots'] = true;
+        }
+
+        return json_encode($config);
+    }
 }